langchain-search-app/requirements.txt
flask
langchain
langchain-community
langchain-groq
python-dotenv
duckduckgo-search
